requirements.txt
aiohttp==3.5.4
galaxy.plugin.api==0.35
pytest==4.2.0
pytest-flakes==4.0.0
pytest-pythonpath==0.7.3
pytest-asyncio==0.10.0
pytest-mock==1.10.3
